Enum kiss_ui::callback::CallbackStatus

pub enum CallbackStatus {
    Default,
    Close,
}

Set this within a callback to tell the framework if it should close or not.

If Callback::close.set() is called within a callback, then when the callback returns, the dialog containing the widget on which the callback was invoked will be closed.

Variants

Default

The default CallbackStatus, does nothing when set.

Close

If this is set within a callback, then when the callback returns the dialog containing the widget on which the callback was invoked will be closed.

Methods

impl CallbackStatus

fn close(&mut self)