Trait kiss_ui::callback::OnClick

pub trait OnClick: Widget {
    fn set_onclick<Cb>(self, on_click: Cb) -> Self where Cb: Callback<Self>;
}

A trait describing a widget that can be clicked, and can notify client code when this occurs.

Required Methods

fn set_onclick<Cb>(self, on_click: Cb) -> Self where Cb: Callback<Self>