Trait kiss_ui::callback::OnValueChange

pub trait OnValueChange: Widget {
    fn set_on_value_changed<Cb>(self, on_value_chaged: Cb) -> Self where Cb: Callback<Self>;
}

A trait describing a widget which has a value that can be changed by the user, and can notify client code when this occurs.

Required Methods

fn set_on_value_changed<Cb>(self, on_value_chaged: Cb) -> Self where Cb: Callback<Self>