Module kiss_ui::container

Assorted types that can contain multiple widgets.

All container types can be nested.

Use the children!{} macro in this crate to convert a heterogeneous list of widgets into a Vec<BaseWidget> for the container constructors.

Structs

Absolute
A container type that makes no effort to arrange its children. Instead, they must be positioned manually.

AbsoluteBuilder
A builder for Absolute, used to create and add children.

Grid
A container widget that lines up its children from left to right, and from top to bottom.

Horizontal
A container widget that lines up its children from left to right.

Vertical
A container widget that lines up its children from top to bottom.

Enums

HAlign
Horizontal alignment setting, used by Vertical and Grid.

Orientation
The behavior of this enum depends on its point of use.

VAlign
Vertical alignment setting, used by Horizontal and Grid.