Struct kiss_ui::container::Absolute

pub struct Absolute(_);

A container type that makes no effort to arrange its children. Instead, they must be positioned manually.

Methods

impl Absolute

fn new<F>(build_fn: F) -> Absolute where F: FnOnce(&mut AbsoluteBuilder)

Create a new absolute container using the given closure, which will be passed a mutable builder instance.

The builder is necessary to ensure that the children have their positions set correctly, as Widget::set_position will not set the attributes that this particular container is expecting.

fn set_child_pos<W: Widget>(x: u32, y: u32, child: W)

Set the position of a child of an Absolute relative to its top-left corner.