Struct kiss_ui::container::AbsoluteBuilder

pub struct AbsoluteBuilder {
    // some fields omitted
}

A builder for Absolute, used to create and add children.

Methods

impl AbsoluteBuilder

fn add_child_at<W: Widget>(&mut self, x: u32, y: u32, child: W) -> &mut Self

Add a child to the Absolute at the given coordinates, relative to the top-left corner of the container.