Module kiss_ui::dialog

KISS-UI top-level dialogs (windows)

Structs

AlertPopupBuilder
A builder for an alert dialog that can show a message and up to 3 buttons for the user's response.

Dialog
A top-level dialog that can create a new native window when shown, and can contain a single widget (which can be a container for many widgets).

Functions

message_popup
Popup a message dialog and block until it is closed, by either the OK button or the exit button.