Struct kiss_ui::dialog::AlertPopupBuilder

pub struct AlertPopupBuilder {
    pub title: String,
    pub message: String,
    pub button1: String,
    pub button2: Option<String>,
    pub button3: Option<String>,
}

A builder for an alert dialog that can show a message and up to 3 buttons for the user's response.

Fields

title

message

button1

button2

button3

Methods

impl AlertPopupBuilder

fn new<T: Into<String>, M: Into<String>, B1: Into<String>>(title: T, message: M, button1: B1) -> AlertPopupBuilder

fn button2<B2: Into<String>>(self, button2: B2) -> Self

Set the text of the second button

fn button3<B3: Into<String>>(self, button3: B3) -> Self

fn popup(self) -> i32

Popup the dialog and block until the user takes an action.

Returns: which button was pressed, or 0 if the dialog was closed.