Trait kiss_ui::image::ImageContainer

pub trait ImageContainer: Widget {
    fn set_image(self, image: Image) -> Self { ... }
    fn get_image(&self) -> Option<Image> { ... }
}

A trait describing an object that can render an image within itself.

Provided Methods

fn set_image(self, image: Image) -> Self

Set the image this widget is to render and return self for method chaining.

fn get_image(&self) -> Option<Image>

Get a copy of the image set on this widget, if any.